Fixing power tile bug, tower stats and specific tower interactions, and enemy set target bug.

+++ Assets/Scripts/Tower Scripts/Projectile Scripts/BulletParent.cs
@@ -4,6 +4,6 @@ using UnityEngine;
 
 public class BulletParent : MonoBehaviour
 {
-    public int _damage;
+    public float _damage;
     public float _bulletSpeed;
 }
